requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.27.0
python-telegram-bot==20.7
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
httpx~=0.25.2
requests>=2.31.0
pydantic-settings>=2.0.0